requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector==2.2.9
redis==3.5.3
requests==2.24.0
rq==1.5.2
soupsieve==2.0.1
urllib3==1.25.11
Werkzeug==1.0.1
